utils.py
import os
from PyPDF2 import PdfReader

def count_pdf_pages(pdf_path):
    try:
        pdf = PdfReader(pdf_path)
        return len(pdf.pages)
    except Exception as e:
        print("Error:",e)
        return None