requirements.txt
cached_property
farm-haystack
fastapi
uvicorn
farm-haystack[weaviate]
pypdf
torch
sentence_transformers
openai
tiktoken
python-dotenv
jinja2